backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "SafetyAdvisor FastAPI Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.3",
    "uvicorn[standard]>=0.34.3",
    "PyPDF2>=3.0.1",
    "pdf2image>=1.17.0",
    "openai>=1.90.0",
    "python-multipart>=0.0.20",
    "mistralai>=1.8.2",
]
